Move settings-related components inside settings directory
