+++ src/components/settings/SettingsForm.svelte
@@ -2,12 +2,12 @@
 import { createEventDispatcher } from 'svelte';
 import isEmpty from 'lodash/isEmpty';
 import Button from '@components/Button.svelte';
-import SettingsFormSidebar from '@components/SettingsFormSidebar.svelte';
-import SettingsFormThemeTab from '@components/SettingsFormThemeTab.svelte';
-import SettingsFormBackgroundTab from '@components/SettingsFormBackgroundTab.svelte';
-import SettingsFormTagsTab from '@components/SettingsFormTagsTab.svelte';
-import SettingsFormShortcutsTab from '@components/SettingsFormShortcutsTab.svelte';
-import SettingsFormMiscellaneousTab from '@components/SettingsFormMiscellaneousTab.svelte';
+import SettingsFormSidebar from '@components/settings/SettingsFormSidebar.svelte';
+import SettingsFormThemeTab from '@components/settings/tabs/SettingsFormThemeTab.svelte';
+import SettingsFormBackgroundTab from '@components/settings/tabs/SettingsFormBackgroundTab.svelte';
+import SettingsFormTagsTab from '@components/settings/tabs/SettingsFormTagsTab.svelte';
+import SettingsFormShortcutsTab from '@components/settings/tabs/SettingsFormShortcutsTab.svelte';
+import SettingsFormMiscellaneousTab from '@components/settings/tabs/SettingsFormMiscellaneousTab.svelte';
 import {
   SETTINGS_THEME,
   SETTINGS_BACKGROUND,
