+++ src/components/settings/fields/SettingsFormBackgroundSourceChoice.svelte
