+++ src/components/settings/fields/SettingsFormColorChoice.svelte
