+++ src/components/settings/tabs/SettingsFormShortcutsTab.svelte
